% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.R
\name{agg}
\alias{agg}
\title{Aggregation}
\usage{
agg(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Simple alias for aggregation functions in data.table
}
\examples{
dt <- data.table(x = c(1,2,3), y = c("a","a","b"))
dt[, agg(avg_x = mean(x)), by = y]
}
