% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_across.R
\name{arrange_across.}
\alias{arrange_across.}
\title{Arrange by a selection of variables}
\usage{
arrange_across.(.df, .cols = everything(), .fns = NULL)
}
\arguments{
\item{.df}{A data.table or data.frame}

\item{.cols}{vector \code{c()} of unquoted column names. \code{tidyselect} compatible.}

\item{.fns}{Function to apply. If \code{desc} it arranges in descending order}
}
\description{
Arrange all rows in either ascending or descending order by a selection of variables.
}
\examples{
test_df <- tidytable(a = c("a", "b", "a"), b = 3:1)

test_df \%>\%
  arrange_across.()

test_df \%>\%
  arrange_across.(a, desc.)
}
