% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dt.R
\name{as_dt}
\alias{as_dt}
\title{Deprecated for as_tidytable()}
\usage{
as_dt(.data)
}
\arguments{
\item{.data}{A data.frame or data.table}
}
\description{
Lifecycle: soft-deprecated

Please use as_tidytable() instead
}
\examples{
example_df <- data.frame(x = 1:10)

example_df \%>\%
  as_tidytable() \%>\%
  dt_mutate(double_x = x * 2)
}
