% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between.R
\name{between.}
\alias{between.}
\title{Do the values from x fall between the left and right bounds?}
\usage{
between.(x, left, right)
}
\arguments{
\item{x}{A numeric vector of values}

\item{left, right}{Boundary values (must be scalars).}
}
\description{
\code{between.()} utilizes \code{data.table::between()} in the background
}
\examples{
between.(1:10, 5, 7)

test_df <- data.frame(x = sample(1:5, 10, replace = TRUE), y = sample(1:5, 10, replace = TRUE))

## Use in a filter
test_df \%>\%
  filter.(between.(x, 1,3))

 ## Use in a mutate
 test_df \%>\%
   mutate.(x_between = between.(x, 2,3),
   x_between_by_two = ifelse.(between.(x, 2,3), x * 2, 0))


}
