% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_across.R
\name{c_across.}
\alias{c_across.}
\title{Combine values from multiple columns}
\usage{
c_across.(cols = everything())
}
\arguments{
\item{cols}{Columns to transform.}
}
\description{
\code{c_across.()} works inside of \code{mutate_rowwise.()}. It uses tidyselect so
you can easily select multiple variables.
}
\examples{
test_df <- data.table(x = runif(6), y = runif(6), z = runif(6))

test_df \%>\%
  mutate_rowwise.(row_mean = mean(c_across.(x:z)))
}
