% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutive_id.R
\name{consecutive_id}
\alias{consecutive_id}
\title{Generate a unique id for consecutive values}
\usage{
consecutive_id(...)
}
\arguments{
\item{...}{Vectors of values}
}
\description{
Generate a unique id for runs of consecutive values
}
\examples{
x <- c(1, 1, 2, 2, 1, 1)
consecutive_id(x)
}
