% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossing.R
\name{crossing}
\alias{crossing}
\title{Create a data.table from all unique combinations of inputs}
\usage{
crossing(..., .name_repair = "check_unique")
}
\arguments{
\item{...}{Variables to get unique combinations of}

\item{.name_repair}{Treatment of problematic names. See \code{?vctrs::vec_as_names} for options/details}
}
\description{
\code{crossing()} is similar to \code{expand_grid()} but de-duplicates and sorts its inputs.
}
\examples{
x <- 1:2
y <- 1:2

crossing(x, y)

crossing(stuff = x, y)
}
