% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative.R
\name{cumall.}
\alias{cumall.}
\alias{cumany.}
\alias{cummean.}
\title{Cumulative versions of any, all, and mean}
\usage{
cumall.(x)

cumany.(x)

cummean.(x)
}
\arguments{
\item{x}{For \code{cumall.()} and \code{cumany.()} a logical vector.
For \code{cummean.()} and integer or numeric vector.}
}
\description{
\code{cumall.()}, \code{cumany.()}, and \code{cummean.()} to work alongside base \code{cumsum()}
}
\examples{
x <- c(1, 3, 5, 2, 2)
cummean.(x)

# `cumall()` and `cumany()` return logicals
cumall.(x < 5)
cumany.(x == 3)

# `cumall()` vs. `cumany()
df <- data.table(
  date = as.Date("2020-01-01") + 0:6,
  balance = c(100, 50, 25, -25, -50, 30, 120)
)

# all rows after first overdraft
df \%>\% filter.(cumany.(balance < 0))

# all rows until first overdraft
df \%>\% filter.(cumall.(!(balance < 0)))
}
