% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc.}
\alias{desc.}
\title{Descending order}
\usage{
desc.(x)
}
\arguments{
\item{x}{Variable to arrange in descending order}
}
\description{
Arrange in descending order. Can be used inside of \code{arrange.()}
}
\examples{
test_df <- data.table(
  a = c(1,2,3),
  b = c(4,5,6),
  c = c("a","a","b"))

test_df \%>\%
  arrange.(c, desc.(a))
}
