% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na.R
\name{drop_na.}
\alias{drop_na.}
\alias{dt_drop_na}
\title{Drop rows containing missing values}
\usage{
drop_na.(.data, ...)

dt_drop_na(.data, ...)
}
\arguments{
\item{.data}{A data.frame or data.table}

\item{...}{Optional: A selection of columns. If empty, all variables are selected.}
}
\description{
Drop rows containing missing values

Supports enhanced selection
}
\examples{
df <- data.table::data.table(
  x = c(1,2,NA),
  y = c("a",NA,"b"))

df \%>\%
  drop_na.()

df \%>\%
  drop_na.(x)

df \%>\%
  drop_na.(is.numeric)
}
