% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{dt_arrange}
\alias{dt_arrange}
\title{Arrange/reorder rows by variables}
\usage{
dt_arrange(.data, ...)
}
\arguments{
\item{.data}{A data.frame or data.table}

\item{...}{Variables to arrange by}
}
\description{
Order rows in ascending or descending order
}
\examples{
example_dt <- data.table::data.table(
  a = c(1,2,3),
  b = c(4,5,6),
  c = c("a","a","b"))

example_dt \%>\%
  dt_arrange(a, -c)
}
