% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{dt_bind_cols}
\alias{dt_bind_cols}
\title{Bind data.tables by row and column}
\usage{
dt_bind_cols(.data, ...)
}
\arguments{
\item{.data}{A data frame to bind}

\item{...}{Data frames to bind}
}
\description{
Simple alias to data.table implementations of \code{rbind()} and \code{cbind()}
}
\examples{

example_dt \%>\%
  dt_bind_rows(another_dt)

example_dt \%>\%
  dt_bind_cols(another_dt)
}
