% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{dt_bind_rows}
\alias{dt_bind_rows}
\title{Bind data.tables by row and column}
\usage{
dt_bind_rows(...)
dt_bind_cols(...)
}
\arguments{
\item{...}{Data frames to combine.}
}
\description{
Simple alias to data.table implementations of \code{rbind()} and \code{cbind()}
}
\examples{

example_dt \%>\%
  dt_bind_rows(another_dt)

example_dt \%>\%
  dt_bind_cols(another_dt)
}
