% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na.R
\name{dt_drop_na}
\alias{dt_drop_na}
\title{Drop rows containing missing values}
\usage{
dt_drop_na(dt_, ...)
}
\arguments{
\item{...}{A selection of columns. If empty, all variables are selected. You can supply bare variable names.}

\item{.data}{A data.frame or data.table}
}
\value{
A data.table
}
\description{
Drop rows containing missing values
}
\examples{
library(data.table)
df <- data.table(x = c(1, 2, NA), y = c("a", NA, "b"))
df \%>\% dt_drop_na()
df \%>\% dt_drop_na(x)
}
