% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{dt_filter}
\alias{dt_filter}
\title{Filter rows on one or more conditions}
\usage{
dt_filter(.data, ...)
}
\arguments{
\item{.data}{A data.frame or data.table}

\item{...}{Conditions to filter by}
}
\description{
Filters a dataset to choose rows where conditions are true.
}
\examples{
example_dt <- data.table::data.table(
  a = c(1,2,3),
  b = c(4,5,6))

example_dt \%>\%
  dt_filter(a >= 2, b >= 4)
}
