% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dummies.R
\name{dt_get_dummies}
\alias{dt_get_dummies}
\title{Convert character and factor columns to dummy variables}
\usage{
dt_get_dummies(
  .data,
  cols = NULL,
  prefix = TRUE,
  prefix_sep = "_",
  drop_first = FALSE,
  dummify_na = TRUE
)
}
\arguments{
\item{.data}{A data.frame or data.table}

\item{cols}{If NULL makes dummy variables for all character & factor columns. Or user can input a single column or a vector of unquoted columns to dummify}

\item{prefix}{TRUE/FALSE - If TRUE, a prefix will be added to new column names}

\item{prefix_sep}{Separator for new column names}

\item{drop_first}{TRUE/FALSE - If TRUE, the first dummy column will be dropped}

\item{dummify_na}{TRUE/FALSE - If TRUE, NAs will also get dummy columns}
}
\description{
Convert character and factor columns to dummy variables

Supports enhanced selection
}
\examples{
test_df <- data.table::data.table(
  col1 = c(letters[1:3], NA),
  col2 = as.factor(c(letters[3:1], NA)),
  var1 = rnorm(7,0,1))

# Automatically does all character/factor columns
test_df \%>\%
  dt_get_dummies()

# Can select one column
test_df \%>\%
  dt_get_dummies(col1)

# Can select one or multiple columns in a vector of unquoted column names
test_df \%>\%
  dt_get_dummies(c(col1, col2))

test_df \%>\%
  dt_get_dummies(prefix_sep = ".", drop_first = TRUE)

test_df \%>\%
  dt_get_dummies(c(col1, col2), dummify_na = FALSE)
}
