% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_split.R
\name{dt_group_split}
\alias{dt_group_split}
\title{Split data frame by groups}
\usage{
dt_group_split(.data, ...)
}
\arguments{
\item{.data}{A data.frame or data.table}

\item{...}{Groups to split by}
}
\description{
Split data frame by groups. Returns a list
}
\examples{
test_df <- data.table::data.table(
  a = 1:5,
  b = 1:5,
  c = c("a","a","a","b","b"),
  d = c("a","a","a","b","b"))

test_df \%>\%
  dt_group_split(c, d)
}
