% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifelse.R
\name{dt_ifelse}
\alias{dt_ifelse}
\title{Vectorized if}
\usage{
dt_ifelse(conditions, true, false)
}
\arguments{
\item{conditions}{Conditions to test on}

\item{true}{Values to return if conditions evaluate to TRUE}

\item{false}{Values to return if conditions evaluate to FALSE}
}
\description{
\code{dt_ifelse()} utilizes \code{data.table}'s \code{fifelse()} in the background, but automatically
converts NAs to their proper type
}
\examples{
x <- c(1:4, 3:2, 1:4)
dt_ifelse(x > 2L, x, x - 1L)

# Can also be used inside of dt_mutate()
test_df <- data.table::data.table(x = x)
test_df \%>\%
  dt_mutate(new_col = dt_ifelse(x > 2L, NA, x - 1L))
}
