% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{dt_left_join}
\alias{dt_left_join}
\alias{dt_inner_join}
\alias{dt_right_join}
\alias{dt_full_join}
\alias{dt_anti_join}
\title{Join two data.tables together}
\usage{
dt_left_join(x, y, by = NULL, suffix = c(".x", ".y"))
dt_inner_join(x, y, by = NULL, suffix = c(".x", ".y"))
dt_right_join(x, y, by = NULL, suffix = c(".x", ".y"))
dt_full_join(x, y, by = NULL, suffix = c(".x", ".y"))

dt_inner_join(x, y, by = NULL, suffix = c(".x", ".y"))

dt_right_join(x, y, by = NULL, suffix = c(".x", ".y"))

dt_full_join(x, y, by = NULL, suffix = c(".x", ".y"))

dt_anti_join(x, y, by = NULL, suffix = c(".x", ".y"))
}
\arguments{
\item{x}{A data.frame or data.table}

\item{y}{A data.frame or data.table}

\item{by}{A character vector of variables to join by. If NULL, the default, the join will do a natural join, using all variables with common names across the two tables.}

\item{suffix}{}
}
\value{
A data.table
}
\description{
Join two data.tables together
}
\examples{
band_members \%>\% dt_left_join(band_instruments)
band_members \%>\% dt_inner_join(band_instruments)
band_members \%>\% dt_right_join(band_instruments)
band_members \%>\% dt_full_join(band_instruments)
}
