% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr.R
\name{dt_map_dfc}
\alias{dt_map_dfc}
\title{Apply a function to each element of a vector}
\usage{
dt_map_dfc(.x, .f, ...)
}
\arguments{
\item{.x}{A list or vector}

\item{.f}{A function}

\item{...}{Other arguments to pass to a function}
}
\value{

}
\description{
The dt_map functions transform their input by applying a function to each element and returning a vector the same length as the input.
\itemize{
\item \code{dt_map()} returns a list
\item \verb{_lgl()}, \verb{_int()}, \verb{_dbl()} and \verb{_chr()} variants return their specified type
\item \verb{_dfr()} & \verb{_dfc()} Return data frame results binded together
}

\emph{Note: the '~' alias will not work with any of the \code{dt_map()} functions}
}
\examples{
1:10 \%>\% map(rnorm, n = 10)
}
