% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_helpers.R
\name{dt_n}
\alias{dt_n}
\title{Summarize helpers}
\usage{
dt_n()
}
\value{
A data.table
}
\description{
These are made to mimic dplyr's \code{summarize()} helpers
}
\examples{
example_dt \%>\%
  dt_summarize(count = dt_n())
}
