% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_longer.R
\name{dt_pivot_longer}
\alias{dt_pivot_longer}
\title{Pivot data from wide to long}
\usage{
dt_pivot_longer(data, cols = NULL, names_to = "name", values_to = "value")
}
\arguments{
\item{data}{The data.table}

\item{cols}{Column selection. If empty uses all columns. Can use -colname to unselect column(s)}

\item{names_to}{Name of the new "names" column. Must be a string.}

\item{values_to}{Name of the new "values" column. Must be a string.}
}
\description{
`pivot_longer()` "lengthens" data, increasing the number of rows and
decreasing the number of columns. The inverse transformation is
`pivot_wider()`
}
\examples{
example_df <- data.table(x = c(1,2,3), y = c(4,5,6), z = c("a", "b", "c"))

example_df \%>\%
  as_dt() \%>\%
  dt_pivot_longer(cols = c(x, y), names_to = "stuff", values_to = "things")
}
