% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_variants.R
\name{dt_rename_all}
\alias{dt_rename_all}
\alias{dt_rename_at}
\alias{dt_rename_across}
\alias{dt_rename_if}
\title{Rename a selection of variables}
\usage{
dt_rename_all(.data, .fun, ...)

dt_rename_at(.data, .vars, .fun, ...)

dt_rename_across(.data, .cols, .fun, ...)

dt_rename_if(.data, .predicate, .fun, ...)
}
\arguments{
\item{.data}{A data.frame or data.table}

\item{.fun}{Function to pass}

\item{...}{Other arguments for the passed function}

\item{.cols}{vector \code{c()} of bare column names for \code{dt_mutate_across()} to use}
}
\value{

}
\description{
These scoped variants of \code{rename()} operate on a selection of variables

There are two variants:
\itemize{
\item \code{dt_rename_all()}
\item \code{dt_rename_across()}: Replaces both \code{mutate_if()} & \code{mutate_at()}
}

Supports enhanced selection
}
\examples{

example_dt <- data.table(x = 1,
                         y = 2,
                         double_x = 2,
                         double_y = 4)

example_dt \%>\% dt_rename_all(str_replace, "x", "stuff")

example_dt \%>\%
  dt_rename_across(c(x, double_x), str_replace, "x", "stuff")

example_dt \%>\%
  dt_rename_across(is.double, function(x) str_replace(x, "x", "stuff"))
}
