% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_variants.R
\name{dt_rename_at}
\alias{dt_rename_at}
\title{Rename a selection of variables}
\usage{
dt_rename_at(.data, .vars, .fun, ...)
}
\arguments{
\item{.data}{A data.frame or data.table}

\item{.vars}{\code{list()} of variables for \code{dt_rename_at()} to use}

\item{.fun}{Function to pass}

\item{...}{Other arguments for the passed function}
}
\value{

}
\description{
These scoped variants of rename() operate on a selection of variables
}
