% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_variants.R
\name{dt_rename_if}
\alias{dt_rename_if}
\title{Rename a selection of variables}
\usage{
dt_rename_if(.data, .predicate, .fun, ...)
}
\arguments{
\item{.data}{A data.frame or data.table}

\item{.predicate}{Predicate to specify columns for \code{dt_rename_if()}}

\item{.fun}{Function to pass}

\item{...}{Other arguments for the passed function}
}
\value{

}
\description{
These scoped variants of `rename()`` operate on a selection of variables
}
