% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na.R
\name{dt_replace_na}
\alias{dt_replace_na}
\title{Replace missing values}
\usage{
dt_replace_na(.col, replace)
}
\arguments{
\item{.col}{A vector}

\item{replace}{A single value used for replacement}
}
\description{
A shortcut to replace NAs inside of a \code{dt_mutate()} call.

Note: This function \emph{does not work} outside of \code{dt_mutate()} like \code{tidyr::replace_na()} does.
}
\examples{
example_dt <- data.table::data.table(
  x = c(1, 2, NA),
  y = c(NA, 1, 2))

example_dt \%>\%
  dt_mutate(x = dt_replace_na(x, 5))
}
