% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{dt_separate}
\alias{dt_separate}
\title{Separate a character column into multiple columns}
\usage{
dt_separate(.data, col, into, sep = "[^[:alnum:]]+", remove = TRUE, ...)
}
\arguments{
\item{.data}{A data.frame or data.table}

\item{col}{The column to split into multiple columns}

\item{into}{New column names to split into. A character vector.}

\item{sep}{Separator to split on. Can be}

\item{remove}{If TRUE, remove the input column from the output data.table}

\item{...}{Further argument to pass to data.table::tstrsplit}
}
\description{
Separates a single column into multiple columns using either regex or a user supplied separator.
}
