% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{dt_slice}
\alias{dt_slice}
\title{Choose rows by position}
\usage{
dt_slice(data, rows = NULL, by = NULL)
}
\arguments{
\item{rows}{Integer row values. Provide either positive values to keep, or negative values to drop. The values provided must be either all positive or all negative.}

\item{by}{A single unquoted column, or a `list()` of columns to group by.}
}
\value{
data.table
}
\description{
Choose rows by their ordinal position in a data.table. Grouped data.tables use the ordinal position within the group.
}
\examples{
library(data.table)

example_dt <- data.table(x = c(1,2,3,4), y = c(4,5,6,7), z = c("a", "a", "a", "b"))

# Select rows
dt_slice(example_dt, 1:10)

# Select rows by group
dt_slice(example_dt, 1, by = z)
}
