% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{dt_summarise}
\alias{dt_summarise}
\title{\code{dplyr} verbs for data.table}
\usage{
dt_summarise(.data, ..., by, keyby)
}
\arguments{
\item{.data}{A data.frame or data.table}

\item{...}{Values passed to \code{gdt} functions}

\item{by}{\code{list()} of bare column names to group by}

\item{keyby}{\code{list()} of bare column names to group by or key by}
}
\value{
A data.table
}
\description{
The core dplyr verbs can be used in \code{gdt}:
\itemize{
\item \code{dt_mutate()}
\item \code{dt_select()}
\item \code{dt_arrange()}
\item \code{dt_filter()}
\item \code{dt_summarize()}
}
}
\examples{
example_dt <- data.table(x = c(1,2,3), y = c(4,5,6), z = c("a","a","b"))

example_dt \%>\%
  dt_select(x, y, z) \%>\%
  dt_mutate(double_x = x * 2,
            double_y = y * 2) \%>\%
  dt_filter(double_x > 0, double_y > 0) \%>\%
  dt_arrange(-double_x) \%>\%
  dt_summarize(avg_x = mean(x), by = z)
}
