% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmute.R
\name{dt_transmute}
\alias{dt_transmute}
\title{Add new variables and drop all others}
\usage{
dt_transmute(.data, ..., by = NULL)
}
\arguments{
\item{.data}{A data.frame or data.table}

\item{...}{Columns to create/modify}

\item{by}{Optional: `list()` of bare column names to group by}
}
\description{
Add new variables and drop all others
}
\examples{
# transmute keeps only the variables you create

mtcars \%>\%
  dt_transmute(displ_l = disp / 61.0237)
}
