% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{dt_unnest}
\alias{dt_unnest}
\title{Unnest a nested data.table}
\usage{
dt_unnest(.data, col, ...)
}
\arguments{
\item{.data}{A nested data.table}

\item{col}{The column to unnest}

\item{...}{Column names to keep}
}
\value{
A data.table
}
\description{
Unnest a nested data.table
}
\examples{
test_df <- data.table(a = 1:10,
                      b = 11:20,
                      c = c(rep("a", 6), rep("b", 4)),
                      d = c(rep("a", 4), rep("b", 6)))
test_df \%>\%
  dt_group_nest(c, d) \%>\%
  dt_unnest(data, c, d)
}
