% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe.R
\name{enframe}
\alias{enframe}
\title{Convert a vector to a data.table/tidytable}
\usage{
enframe(x, name = "name", value = "value")
}
\arguments{
\item{x}{A vector}

\item{name}{Name of the column that stores the names. If \code{name = NULL},
a one-column tidytable will be returned.}

\item{value}{Name of the column that stores the values.}
}
\description{
Converts named and unnamed vectors to a data.table/tidytable.
}
\examples{
vec <- 1:3
names(vec) <- letters[1:3]

enframe(vec)
}
