% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlist_dots.R
\name{enlist_dots}
\alias{enlist_dots}
\title{Enlist dots input by the user}
\usage{
enlist_dots(...)
}
\arguments{
\item{...}{User inputs dots}
}
\description{
Enlist dots input by the user. Equivalent to \code{rlang::enexprs()}
}
\examples{

new_fn <- function(.data, ...) {
  dots <- enlist_dots(...)
}
}
