% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_grid.R
\name{expand_grid.}
\alias{expand_grid.}
\title{Create a data.table from all combinations of inputs}
\usage{
expand_grid.(..., .name_repair = "check_unique")
}
\arguments{
\item{...}{Variables to get combinations of}

\item{.name_repair}{Treatment of problematic names. See \code{?vctrs::vec_as_names} for options/details}
}
\description{
Create a data.table from all combinations of inputs
}
\examples{
x <- 1:2
y <- 1:2

expand_grid.(x, y)

expand_grid.(stuff = x, y)
}
