% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-last.R
\name{first.}
\alias{first.}
\alias{last.}
\title{Extract the first or last value from a vector}
\usage{
first.(x)

last.(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Extract the first or last value from a vector.
}
\examples{
vec <- letters

first.(vec)
last.(vec)
}
