% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread-fwrite.R
\name{fread.}
\alias{fread.}
\alias{fwrite.}
\title{Read/write files}
\usage{
fread.(...)

fwrite.(...)
}
\arguments{
\item{...}{Arguments passed on to \code{data.table::fread} or \code{data.table::fwrite}.
See \code{?data.table::fread} or \code{?data.table::fwrite} for details}
}
\description{
\code{fread.()} is a simple wrapper around \code{data.table::fread()} that returns a tidytable
instead of a data.table.

\code{fwrite.()} is a simple wrapper around \code{data.table::fwrite()}.
}
\examples{
fake_csv <- "A,B
             1,2
             3,4"

fread.(fake_csv)
}
