% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread-fwrite.R
\name{fwrite.}
\alias{fwrite.}
\title{Read/write files}
\usage{
fwrite.(...)
}
\arguments{
\item{...}{Arguments passed on to \code{data.table::fread}}
}
\description{
\code{fread()} is a simple wrapper around \code{data.table::fread()} that returns a tidytable
instead of a data.table.
}
\examples{
fake_csv <- "A,B
             1,2
             3,4"

fread(fake_csv)
}
\keyword{internal}
