% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{notin operator}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{vector or NULL}

\item{y}{vector or NULL}
}
\description{
notin operator
}
\examples{
c(1,3,11) \%notin\% 1:10
}
