% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_cols.R
\name{group_cols}
\alias{group_cols}
\title{Selection helper for grouping columns}
\usage{
group_cols()
}
\description{
Selection helper for grouping columns
}
\examples{
df <- tidytable(
  x = c("a", "b", "c"),
  y = 1:3,
  z = 1:3
)

df \%>\%
  group_by(x) \%>\%
  select(group_cols(), y)
}
