% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_vars.}
\alias{group_vars.}
\title{Get the grouping variables}
\usage{
group_vars.(x)
}
\arguments{
\item{x}{A grouped tidytable}
}
\description{
Get the grouping variables
}
\examples{
df <- data.table(
  a = 1:3,
  b = 4:6,
  c = c("a", "a", "b"),
  d = c("a", "a", "b")
)

df \%>\%
  group_by(c, d) \%>\%
  group_vars()
}
\keyword{internal}
