% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_all-if_any.R
\name{if_all.}
\alias{if_all.}
\alias{if_any.}
\title{Create conditions on a selection of columns}
\usage{
if_all.(.cols = everything(), .fns = NULL, ...)

if_any.(.cols = everything(), .fns = NULL, ...)
}
\arguments{
\item{.cols}{Selection of columns}

\item{.fns}{Function to create filter conditions}

\item{...}{Other arguments passed to the function}
}
\description{
Helpers to apply a filter across a selection of columns.
}
\examples{
iris \%>\%
  filter.(if_any.(ends_with("Width"), ~ .x > 4))

iris \%>\%
  filter.(if_all.(ends_with("Width"), ~ .x > 2))
}
