% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_gc.R
\name{inv_gc}
\alias{inv_gc}
\title{Run invisible garbage collection}
\usage{
inv_gc(df = NULL)
}
\value{

}
\description{
Run garbage collection without the gc() output. Can also be run in the middle of a long pipe chain. Useful for large datasets.
}
\examples{
inv_gc()

df \%>\%
  inv_gc() \%>\%
  select(col1, col2)
}
