% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{is_grouped_df.}
\alias{is_grouped_df.}
\title{Check if the tidytable is grouped}
\usage{
is_grouped_df.(x)
}
\arguments{
\item{x}{An object}
}
\description{
Check if the tidytable is grouped
}
\examples{
df <- data.table(
  a = 1:3,
  b = c("a", "a", "b")
)

df \%>\%
  group_by(b) \%>\%
  is_grouped_df()
}
\keyword{internal}
