% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.tidytable}
\alias{knit_print.tidytable}
\title{knit_print method for tidytables}
\usage{
knit_print.tidytable(x, ..., n = NULL, width = NULL, n_extra = NULL)
}
\description{
knit_print method for tidytables
}
\keyword{internal}
