% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{left_join.}
\alias{left_join.}
\alias{inner_join.}
\alias{right_join.}
\alias{full_join.}
\alias{anti_join.}
\title{Join two data.tables together}
\usage{
left_join.(x, y, by = NULL)

inner_join.(x, y, by = NULL)

right_join.(x, y, by = NULL)

full_join.(x, y, by = NULL, suffix = c(".x", ".y"))

anti_join.(x, y, by = NULL)
}
\arguments{
\item{x}{A data.frame or data.table}

\item{y}{A data.frame or data.table}

\item{by}{A character vector of variables to join by. If NULL, the default, the join will do a natural join, using all variables with common names across the two tables.}

\item{suffix}{Append created for duplicated column names when using \code{full_join.()}}
}
\value{
A data.table
}
\description{
Join two data.tables together
}
\examples{
df1 <- data.table(x = c("a","a","a","b","b"), y = 1:5)
df2 <- data.table(x = c("a","b"), z = 1:2)

df1 \%>\% left_join.(df2)
df1 \%>\% inner_join.(df2)
df1 \%>\% right_join.(df2)
df1 \%>\% full_join.(df2)
df1 \%>\% anti_join.(df2)
}
