% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/let.R
\name{let_if}
\alias{let_if}
\title{Let}
\usage{
let_if(.data, .predicate, .fun, ...) \cr
let_at(.data, .vars, .fun, ...)
}
\arguments{
\item{.data}{The data.table}

\item{.predicate}{Predicate to specify columns for \code{let_if()}}

\item{.fun}{Function to pass}

\item{...}{Other arguments for the passed function}

\item{.vars}{Variables passed to \code{let_at()}}
}
\value{

}
\description{
Add columns. There are two variants of \code{let()}:
\itemize{
\item \code{let_if()}: Equivalent to \code{dplyr::mutate_if()}
\item \code{let_at()}: Equivalent to \code{dplyr::mutate_at()}
}
}
\examples{
example_dt <- data.table(x = c(1,2,3), y = c(4,5,6), z = c("a", "a", "b"))

example_dt[, let(double_x = x * 2)]

example_dt \%>\%
  let_if(is.double, as.character)

example_dt \%>\%
  let_at(vars(x, y), function(.x) .x * 2)
}
