% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_distinct.R
\name{n_distinct.}
\alias{n_distinct.}
\title{Count the number of unique values in a vector}
\usage{
n_distinct.(..., na.rm = FALSE)
}
\arguments{
\item{...}{vectors of values}

\item{na.rm}{If \code{TRUE} missing values don't count}
}
\description{
This is a faster version of \code{length(unique(x))} that calls \code{data.table::uniqueN()}.
}
\examples{
x <- sample(1:10, 1e5, rep = TRUE)
n_distinct(x)
}
\keyword{internal}
