% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_if.R
\name{na_if.}
\alias{na_if.}
\title{Convert values to NA}
\usage{
na_if.(x, y)
}
\arguments{
\item{x}{An R object}

\item{y}{Value(s) to replace with \code{NA}}
}
\description{
Convert values to \code{NA}.
}
\examples{
vec <- 1:3
na_if.(vec, 3)
}
