% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-last-nth.R
\name{nth.}
\alias{nth.}
\title{Extract the first, last, or nth value from a vector}
\usage{
nth.(x, n, default = NULL, na_rm = FALSE)
}
\arguments{
\item{x}{A vector}

\item{n}{For \code{nth()}, a number specifying the position to grab.}

\item{default}{The default value if the value doesn't exist.}

\item{na_rm}{If \code{TRUE} ignores missing values.}
}
\description{
Extract the first, last, or nth value from a vector.

Note: These are simple wrappers around \code{vctrs::vec_slice()}.
}
\examples{
vec <- letters

first(vec)
last(vec)
nth(vec, 4)
}
\keyword{internal}
