% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.R
\name{pick}
\alias{pick}
\title{Selection version of \code{across()}}
\usage{
pick(...)
}
\arguments{
\item{...}{Columns to select. Tidyselect compatible.}
}
\description{
Select a subset of columns from within functions like \code{mutate()}, \code{summarize()}, or \code{filter()}.
}
\examples{
df <- tidytable(
  x = 1:3,
  y = 4:6,
  z = c("a", "a", "b")
)

df \%>\%
  mutate(row_sum = rowSums(pick(x, y)))
}
