% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{row_number}
\alias{row_number}
\alias{min_rank}
\alias{dense_rank}
\alias{percent_rank}
\alias{cume_dist}
\title{Ranking functions}
\usage{
row_number(x)

min_rank(x)

dense_rank(x)

percent_rank(x)

cume_dist(x)
}
\arguments{
\item{x}{A vector to rank}
}
\description{
Ranking functions:
\itemize{
\item \code{row_number()}: Gives other row number if empty.
Equivalent to \code{frank(ties.method = "first")} if provided a vector.
\item \code{min_rank()}: Equivalent to \code{frank(ties.method = "min")}
\item \code{dense_rank()}: Equivalent to \code{frank(ties.method = "dense")}
\item \code{percent_rank()}: Ranks by percentage from 0 to 1
\item \code{cume_dist()}: Cumulative distribution
}
}
\examples{
df <- data.table(x = rep(1, 3), y = c("a", "a", "b"))

df \%>\%
  mutate(row = row_number())
}
