% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Recode values}
\usage{
recode(.x, ..., .default = NULL, .missing = NULL)
}
\arguments{
\item{.x}{A vector}

\item{...}{A series of \code{old = new} pairs specifying the new values}

\item{.default}{The default value if all conditions evaluate to \code{FALSE}}

\item{.missing}{What missing values should be replaced with}
}
\description{
\emph{superseded}

\code{recode()} has been superseded by \code{case_match()}.

Replace old values of a vector with new values.
}
\examples{
char_vec <- c("a", "b", "c")
recode(char_vec, a = "Apple", b = "Banana")

num_vec <- 1:3
recode(num_vec, `1` = 10, `2` = 25, .default = 100)
}
\keyword{internal}
