% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{starts_with}
\alias{contains}
\alias{ends_with}
\alias{everything}
\alias{any_of}
\alias{all_of}
\alias{matches}
\alias{num_range}
\alias{last_col}
\alias{where}
\alias{data.table}
\alias{fwrite}
\alias{getDTthreads}
\alias{setDTthreads}
\alias{\%between\%}
\alias{\%like\%}
\alias{\%chin\%}
\alias{enexpr}
\alias{enexprs}
\alias{enquo}
\alias{enquos}
\alias{expr}
\alias{exprs}
\alias{quo}
\alias{quos}
\alias{sym}
\alias{syms}
\alias{glimpse}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table:\%between\%]{\%between\%}}, \code{\link[data.table:\%chin\%]{\%chin\%}}, \code{\link[data.table:\%like\%]{\%like\%}}, \code{\link[data.table:data.table]{data.table()}}, \code{\link[data.table:fwrite]{fwrite()}}, \code{\link[data.table:getDTthreads]{getDTthreads()}}, \code{\link[data.table:setDTthreads]{setDTthreads()}}}

  \item{pillar}{\code{\link[pillar:glimpse]{glimpse()}}}

  \item{rlang}{\code{\link[rlang:enexpr]{enexpr()}}, \code{\link[rlang:enexprs]{enexprs()}}, \code{\link[rlang:enquo]{enquo()}}, \code{\link[rlang:enquos]{enquos()}}, \code{\link[rlang:expr]{expr()}}, \code{\link[rlang:exprs]{exprs()}}, \code{\link[rlang:quo]{quo()}}, \code{\link[rlang:quos]{quos()}}, \code{\link[rlang:sym]{sym()}}, \code{\link[rlang:syms]{syms()}}}

  \item{tidyselect}{\code{\link[tidyselect:all_of]{all_of()}}, \code{\link[tidyselect:any_of]{any_of()}}, \code{\link[tidyselect:contains]{contains()}}, \code{\link[tidyselect:ends_with]{ends_with()}}, \code{\link[tidyselect:everything]{everything()}}, \code{\link[tidyselect:last_col]{last_col()}}, \code{\link[tidyselect:matches]{matches()}}, \code{\link[tidyselect:num_range]{num_range()}}, \code{\link[tidyselect:starts_with]{starts_with()}}, \code{\link[tidyselect:where]{where()}}}
}}

