% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename.}
\alias{rename.}
\alias{dt_rename}
\title{Rename variables by name}
\usage{
rename.(.data, ...)

dt_rename(.data, ...)
}
\arguments{
\item{.data}{A data.frame or data.table}

\item{...}{Rename expression like dplyr::rename()}
}
\description{
Rename variables from a data.table.
}
\examples{
dt <- data.table::data.table(x = c(1,2,3), y = c(4,5,6))

dt \%>\%
  rename.(new_x = x,
          new_y = y)
}
