% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\alias{rename.}
\title{Rename variables by name}
\usage{
rename(.df, ...)

rename.(.df, ...)
}
\arguments{
\item{.df}{A data.frame or data.table}

\item{...}{}
}
\description{
Rename variables from a data.table.
}
\examples{
df <- data.table(x = 1:3, y = 4:6)

df \%>\%
  rename(new_x = x,
         new_y = y)
}
\keyword{internal}
