% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_number.R
\name{row_number.}
\alias{row_number.}
\alias{dt_row_number}
\title{Return row number}
\usage{
row_number.()

dt_row_number()
}
\description{
This function is designed to work inside of \code{mutate.()}
}
\examples{
test_df <- data.table(x = c(1,1,1))

test_df \%>\%
  mutate.(row = row_number.())
}
