% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_number.R
\name{row_number}
\alias{row_number}
\title{Return row number}
\usage{
row_number()
}
\description{
Returns row number. This function is designed to work inside of \code{mutate()}
}
\examples{
df <- data.table(x = rep(1, 3), y = c("a", "a", "b"))

df \%>\%
  mutate(row = row_number())
}
