% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_across.R
\name{summarize_across.}
\alias{summarize_across.}
\title{Summarize multiple columns}
\usage{
summarize_across.(.df, .cols = everything(), .fns, ..., by = NULL)
}
\arguments{
\item{.df}{A data.frame or data.table}

\item{.cols}{vector \code{c()} of unquoted column names. \code{tidyselect} compatible.}

\item{.fns}{Functions to pass. Can pass a list of functions.}

\item{...}{Other arguments for the passed function}

\item{by}{Columns to group by}
}
\description{
Summarize multiple columns simultaneously
}
\examples{
test_df <- tidytable(a = 1:3, b = 4:6, z = c("a", "a", "b"))

# Can pass a function and arguments separately
test_df \%>\%
  summarize_across.(c(a, b), mean, na.rm = TRUE)

# Or can use a purrr style formula
test_df \%>\%
  summarize_across.(c(a, b), ~ mean(.x, na.rm = TRUE))

# Multiple functions can be passed in a list
test_df \%>\%
  summarize_across.(c(a, b), list(avg = mean, length), by = z)

test_df \%>\%
  summarize_across.(where(is.numeric), mean, na.rm = TRUE)
}
