% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytable.R
\name{tidytable}
\alias{tidytable}
\title{Build a data.table}
\usage{
tidytable(...)
}
\arguments{
\item{...}{Arguments passed to \code{data.table()}}
}
\description{
\code{tidytable()} constructs a data.table, but one with nice printing features.
As such it can be effectively used as a data.table would be used.
}
\examples{
`tidytable(x = c(1,2,3), y = c(4,5,6))`
}
