% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble.R
\name{tribble}
\alias{tribble}
\title{Rowwise tidytable creation}
\usage{
tribble(...)
}
\arguments{
\item{...}{Column names as formulas, values below. See example.}
}
\description{
Create a tidytable using a rowwise setup.
}
\examples{
tribble(
  ~ x, ~ y,
  "a", 1,
  "b", 2,
  "c", 3
)
}
