% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncount.R
\name{uncount.}
\alias{uncount.}
\title{Uncount a data.table}
\usage{
uncount.(.df, weights, .remove = TRUE, .id = NULL)
}
\arguments{
\item{.df}{A data.frame or data.table}

\item{weights}{A column containing the weights to uncount by}

\item{.remove}{If TRUE removes the selected \code{weights} column}

\item{.id}{A string name for a new column containing a unique identifier for the newly uncounted rows.}
}
\description{
Uncount a data.table
}
\examples{
df <- data.table(x = c("a", "b"), n = c(1, 2))

uncount(df, n)

uncount(df, n, .id = "id")
}
\keyword{internal}
