% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tidytable.R
\name{as_dt}
\alias{as_dt}
\title{Deprecated tidytable conversion}
\usage{
as_dt(x)
}
\arguments{
\item{x}{An R object}
}
\description{
Deprecated way to convert an object to a tidytable. Please use \code{as_tidytable()}.
}
\examples{
test_df <- data.frame(x = -2:2, c(rep("a", 3), rep("b", 2)))

test_df \%>\%
  as_tidytable()
}
