% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tidytable.R
\name{as_tidytable}
\alias{as_tidytable}
\alias{as_dt}
\title{Coerce an object to a data.table/tidytable}
\usage{
as_tidytable(x)

as_dt(x)
}
\arguments{
\item{x}{An R object}
}
\description{
A tidytable object is simply a data.table with nice printing features.

Note that all tidytable functions automatically convert data.frames & data.tables to tidytables in the background.
As such this function will rarely need to be used by the user.
}
\examples{
data.frame(x = 1:3) \%>\%
  as_tidytable()

data.frame(x = 1:3) \%>\%
  as_dt()
}
