% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tidytable.R
\name{as_tidytable}
\alias{as_tidytable}
\title{Coerce an object to a data.table/tidytable}
\usage{
as_tidytable(x)
}
\arguments{
\item{x}{An R object}
}
\description{
A tidytable object is simply a data.table with nice printing features.
As such it can be used exactly like a data.table would be used.

Note that all tidytable functions automatically convert data.frames & data.tables to tidytables in the background.
As such this function will rarely need to be used by the user.
}
\examples{
test_df <- data.frame(x = -2:2, y = c(rep("a", 3), rep("b", 2)))

test_df \%>\%
  as_tidytable()
}
