% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc.}
\alias{desc.}
\title{Descending order}
\usage{
desc.(x)
}
\arguments{
\item{x}{Variable to arrange in descending order}
}
\description{
Arrange in descending order. Can be used inside of \code{arrange()}
}
\examples{
df <- data.table(
  a = 1:3,
  b = 4:6,
  c = c("a", "a", "b")
)

df \%>\%
  arrange(c, desc(a))
}
\keyword{internal}
