% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na.R
\name{drop_na.}
\alias{drop_na.}
\title{Drop rows containing missing values}
\usage{
drop_na.(.df, ...)
}
\arguments{
\item{.df}{A data.frame or data.table}

\item{...}{Optional: A selection of columns. If empty, all variables are selected.
\code{tidyselect} compatible.}
}
\description{
Drop rows containing missing values
}
\examples{
df <- data.table(
  x = c(1, 2, NA),
  y = c("a", NA, "b")
)

df \%>\%
  drop_na.()

df \%>\%
  drop_na.(x)

df \%>\%
  drop_na.(where(is.numeric))
}
