% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{dt}
\alias{dt}
\title{Pipeable data.table call}
\usage{
dt(.data, ...)
}
\arguments{
\item{.data}{A data.frame or data.table}

\item{...}{Arguments passed to data.table call. See ?data.table:::`[.data.table`}
}
\description{
Pipeable data.table call

Note: This function does not use data.table's modify-by-reference
}
\examples{
example_dt <- tidytable(
  x = c(1,2,3),
  y = c(4,5,6),
  z = c("a", "a", "b"))

example_dt \%>\%
  dt(, ':='(double_x = x * 2)) \%>\%
  dt(order(-double_x))
}
