% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{dt_bind_rows}
\alias{dt_bind_rows}
\alias{dt_bind_cols}
\title{Bind data.tables by row and column}
\usage{
dt_bind_rows(..., .id = NULL)

dt_bind_cols(...)
}
\arguments{
\item{...}{data.tables or data.frames to bind}

\item{.id}{If TRUE, an integer column is made as a group id}
}
\description{
Bind rows or columns together.
}
\examples{
df1 <- data.table::data.table(x = c(1,2,3), y = c(3,4,5))
df2 <- data.table::data.table(x = c(1,2,3), y = c(3,4,5))

df1 \%>\%
  dt_bind_rows(df2)

dt_bind_rows(list(df1, df2))

df1 \%>\%
  dt_bind_cols(df2)

dt_bind_cols(list(df1, df2))
}
