% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_variants.R
\name{dt_mutate_if}
\alias{dt_mutate_if}
\title{Mutate}
\usage{
dt_mutate_if(.data, .predicate, .fun, ...)
dt_mutate_at(.data, .vars, .fun, ...)
dt_mutate_all(.data, .fun, ...)
}
\arguments{
\item{.data}{A data.frame or data.table}

\item{.predicate}{Predicate to specify columns for \code{dt_mutate_if()}}

\item{.fun}{Function to pass}

\item{...}{Other arguments for the passed function}

\item{.vars}{vector \code{c()} of bare column names for \code{dt_rename_at()} to use}
}
\value{
A data.table
}
\description{
Edit existing columns. There are three variants:
\itemize{
\item \code{dt_mutate_if()}
\item \code{dt_mutate_at()}
\item \code{dt_mutate_all()}
}
}
\examples{
example_dt <- data.table(x = c(1,2,3), y = c(4,5,6), z = c("a", "a", "b"))

example_dt \%>\%
  dt_mutate_if(is.double, as.character)

example_dt \%>\%
  dt_mutate_at(list(x, y), function(.x) .x * 2)
}
