% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_variants.R
\name{dt_mutate_if}
\alias{dt_mutate_if}
\title{Mutate}
\usage{
dt_mutate_if(.data, .predicate, .fun, ...)
dt_mutate_at(.data, .vars, .fun, ...)
}
\arguments{
\item{.data}{The data.table}

\item{.predicate}{Predicate to specify columns for `dt_mutate_if()`}

\item{.fun}{Function to pass}

\item{...}{Other arguments for the passed function}

\item{.vars}{`list()` of variables for `dt_mutate_at()` to use}
}
\value{

}
\description{
Edit existing columns. There are two variants:

* `dt_mutate_if()`
* `dt_mutate_at()`
}
\examples{
example_dt <- data.table(x = c(1,2,3), y = c(4,5,6), z = c("a", "a", "b"))

example_dt \%>\%
  dt_mutate_if(is.double, as.character)

example_dt \%>\%
  dt_mutate_at(list(x, y), function(.x) .x * 2)
}
