% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_variants.R
\name{dt_mutate_if}
\alias{dt_mutate_if}
\alias{dt_mutate_at}
\alias{dt_mutate_across}
\alias{dt_mutate_all}
\title{Mutate}
\usage{


dt_mutate_at(.data, .cols, .fun, ...)

dt_mutate_across(.data, .cols, .fun, ...)

dt_mutate_all(.data, .fun, ...)
}
\arguments{
\item{.data}{A data.frame or data.table}

\item{.fun}{Function to pass}

\item{...}{Other arguments for the passed function}

\item{.cols}{vector \code{c()} of bare column names for \code{dt_mutate_across()} to use}
}
\value{
A data.table
}
\description{
Edit existing columns. Supports enhanced selection.

There are two variants:
\itemize{
\item \code{dt_mutate_all()}
\item \code{dt_mutate_across()}: Replaces both \code{mutate_if()} & \code{mutate_at()}
}
}
\examples{
example_dt <- data.table(x = c(1,1,1),
                         y = c(2,2,2),
                         z = c("a", "a", "b"))

example_dt \%>\%
  as_dt() \%>\%
  dt_mutate_across(is.numeric, as.character)

example_dt \%>\%
  as_dt() \%>\%
  dt_mutate_across(c(x, y), ~ .x * 2)

example_dt \%>\%
  as_dt() \%>\%
  dt_mutate_across(c(x, y), list(new = ~ .x * 2))
}
