% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_wider.R
\name{dt_pivot_wider}
\alias{dt_pivot_wider}
\title{Pivot data from long to wide}
\usage{
dt_pivot_wider(
  data,
  id_cols = NULL,
  names_from = name,
  names_sep = "_",
  values_from = value,
  drop = FALSE
)
}
\arguments{
\item{id_cols}{A set of columns that uniquely identifies each observation.
Defaults to all columns in `data` except for the columns specified in
`names_from` and `values_from`. Typically used when you have additional
variables that is directly related.}

\item{names_from, values_from}{A pair of arguments describing which column
(or columns) to get the name of the output column (`name_from`), and
which column (or columns) to get the cell values from (`values_from`).}

\item{drop}{will cast by including all missing combinations.
c(FALSE, TRUE) will only include all missing combinations of formula LHS;
c(TRUE, FALSE) will only include all missing combinations of formula RHS. See Examples.}
}
\description{
`pivot_wider()` "widens" data, increasing the number of columns and
decreasing the number of rows. The inverse transformation is
[pivot_longer()].
}
\examples{
library(data.table)

example_dt <- data.table(z = rep(c("a", "b", "c"), 2),
                         stuff = c(rep("x", 3), rep("y", 3)),
                         things = 1:6)

example_dt \%>\%
  dt_pivot_wider(names_from = stuff, values_from = things)
}
