% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{dt_pull}
\alias{dt_pull}
\title{Pull out a single variable}
\usage{
dt_pull(.data, var = NULL)
}
\arguments{
\item{.data}{A data.frame or data.table}

\item{var}{The column to pull from the data.table}
}
\value{
A vector
}
\description{
Pull a single variable from a data.table as a vector.
}
\examples{
example_dt <- data.table::data.table(
  x = c(1,2,3),
  y = c(4,5,6))

example_dt \%>\%
  dt_pull(y)
}
