% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{dt_rename}
\alias{dt_rename}
\title{Title}
\usage{
dt_rename(.data, ...)
}
\arguments{
\item{...}{Rename expression like dplyr::rename()}
}
\value{

}
\description{
Title
}
\examples{
dt <- data.table(x = c(1,2,3), y = c(4,5,6))
dt \%>\%
  rename(new_x = x,
         new_y = y)

}
