% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_variants.R
\name{dt_rename_all}
\alias{dt_rename_all}
\title{Rename a selection of variables}
\usage{
dt_rename_all(.data, .fun, ...)
dt_rename_at(.data, .vars, .fun, ...)
dt_rename_if(.data, .predicate, .fun, ...)
}
\arguments{
\item{.data}{A data.frame or data.table}

\item{.fun}{Function to pass}

\item{...}{Other arguments for the passed function}

\item{.predicate}{Predicate to specify columns for \code{dt_mutate_if()}}

\item{.vars}{\code{list()} of variables for \code{dt_mutate_at()} to use}
}
\value{

}
\description{
These scoped variants of rename() operate on a selection of variables
}
