% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_variants.R
\name{dt_select_if}
\alias{dt_select_if}
\title{Select variables}
\usage{
dt_select_if(.data, .predicate)
}
\arguments{
\item{.data}{A data.table}

\item{.predicate}{Predicate to specify columns for `dt_select_if()`}
}
\value{
A data.table
}
\description{
Scoped variants of `select()`
}
\examples{
example_dt <- data.table::data.table(x = 1, y = 2, z = "a")

example_dt \%>\% dt_select_if(is.double)
}
