% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_helpers.R
\name{dt_starts_with}
\alias{dt_starts_with}
\title{Select helpers}
\usage{
dt_starts_with(match)
}
\arguments{
\item{match}{}
}
\value{

}
\description{
These functions allow you to select variables based on their names.
\itemize{
\item \code{dt_starts_with()}: Starts with a prefix
\item \code{dt_starts_with()}: Ends with a suffix
\item \code{dt_contains()}: Contains a literal string
\item \code{dt_everything()}: Matches all variables
}
}
