% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlist_dots.R
\name{enlist_dots}
\alias{enlist_dots}
\title{Enlist Dots}
\usage{
enlist_dots(...)
}
\arguments{
\item{...}{}
}
\value{
list Returns a list for delayed evaluation in a function
}
\description{
Helper function to capture ... inputs from the user.
}
