% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{notin operator}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{vector or NULL}

\item{y}{vector or NULL}
}
\description{
notin operator
}
\examples{
test_df <- tidytable(x = 1:4, y = 1:4)

test_df \%>\%
  filter.(x \%notin\% c(2, 4))
}
