% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in-notin.R
\name{\%in\%}
\alias{\%in\%}
\alias{\%notin\%}
\title{\code{\%in\%} and \verb{\%notin\%} operators}
\usage{
x \%in\% y

x \%notin\% y
}
\arguments{
\item{x}{vector or NULL}

\item{y}{vector or NULL}
}
\description{
Check with values in a vector are in or not in another vector.

Built using \verb{data.table::\%chin\%} and \code{vctrs::vec_in} for performance.
}
\examples{
df <- tidytable(x = 1:4, y = 1:4)

df \%>\%
  filter(x \%in\% c(2, 4))

df \%>\%
  filter(x \%notin\% c(2, 4))
}
