% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tidytable.R
\name{is_tidytable}
\alias{is_tidytable}
\title{Test if the object is a tidytable}
\usage{
is_tidytable(x)
}
\arguments{
\item{x}{An object}
}
\description{
This function returns TRUE for tidytables or subclasses of tidytables, and FALSE for all other objects.
}
\examples{
dt <- data.table::data.table(x = 1)

is_tidytable(dt)

df <- tidytable(x = 1)

is_tidytable(df)
}
