% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{n.}
\alias{n.}
\alias{dt_n}
\title{Number of observations in each group}
\usage{
n.()

dt_n()
}
\description{
Helper function that can be used to find counts by group.

Can be used inside `summarize.()`, `mutate.()`, & `filter.()`
}
\examples{
test_df <- data.table(
  x = c(1,2,3),
  y = c(4,5,6),
  z = c("a","a","b"))

test_df \%>\%
  summarize.(count = n.(),
             .by = z)

test_df \%>\%
  mutate.(count = n.())
}
