% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{n.}
\alias{n.}
\title{Number of observations in each group}
\usage{
n.()
}
\description{
Helper function that can be used to find counts by group.

Can be used inside \code{summarize()}, \code{mutate()}, & \code{filter()}
}
\examples{
df <- data.table(
  x = 1:3,
  y = 4:6,
  z = c("a","a","b")
 )

df \%>\%
  summarize(count = n(), .by = z)
}
\keyword{internal}
