% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_tidytable.R
\name{new_tidytable}
\alias{new_tidytable}
\title{Create a tidytable from a list}
\usage{
new_tidytable(x = list())
}
\arguments{
\item{x}{A named list of equal-length vectors. The lengths are not checked; it is the responsibility
of the caller to make sure they are equal.}
}
\description{
Create a tidytable from a list
}
\examples{
l <- list(x = 1:3, y = c("a", "a", "b"))

new_tidytable(l)
}
