% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull.}
\alias{pull.}
\title{Pull out a single variable}
\usage{
pull.(.df, var = NULL)
}
\arguments{
\item{.df}{A data.frame or data.table}

\item{var}{The column to pull from the data.table. If NULL, pulls the last column.}
}
\description{
Pull a single variable from a data.table as a vector.
}
\examples{
test_df <- data.table(
  x = c(1,2,3),
  y = c(4,5,6))

test_df \%>\%
  pull.(y)
}
