% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytable_package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{starts_with}
\alias{contains}
\alias{ends_with}
\alias{everything}
\alias{any_of}
\alias{all_of}
\alias{matches}
\alias{num_range}
\alias{last_col}
\alias{data.table}
\alias{setDTthreads}
\alias{enquo}
\alias{enquos}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{data.table}}, \code{\link[data.table]{setDTthreads}}}

  \item{rlang}{\code{\link[rlang]{enquo}}, \code{\link[rlang]{enquos}}}

  \item{tidyselect}{\code{\link[tidyselect]{all_of}}, \code{\link[tidyselect]{any_of}}, \code{\link[tidyselect]{contains}}, \code{\link[tidyselect]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect]{last_col}}, \code{\link[tidyselect]{matches}}, \code{\link[tidyselect]{num_range}}, \code{\link[tidyselect]{starts_with}}}
}}

