% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytable-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{starts_with}
\alias{contains}
\alias{ends_with}
\alias{everything}
\alias{any_of}
\alias{all_of}
\alias{matches}
\alias{num_range}
\alias{last_col}
\alias{\%between\%}
\alias{data.table}
\alias{setDTthreads}
\alias{enquo}
\alias{enquos}
\alias{glimpse}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table:between]{\%between\%}}, \code{\link[data.table]{data.table}}, \code{\link[data.table:openmp-utils]{setDTthreads}}}

  \item{rlang}{\code{\link[rlang:nse-defuse]{enquo}}, \code{\link[rlang:nse-defuse]{enquos}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}}

  \item{tidyselect}{\code{\link[tidyselect]{all_of}}, \code{\link[tidyselect:all_of]{any_of}}, \code{\link[tidyselect:starts_with]{contains}}, \code{\link[tidyselect:starts_with]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect:everything]{last_col}}, \code{\link[tidyselect:starts_with]{matches}}, \code{\link[tidyselect:starts_with]{num_range}}, \code{\link[tidyselect]{starts_with}}}
}}

