% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytable.R
\name{tidytable}
\alias{tidytable}
\title{Build a data.table/tidytable}
\usage{
tidytable(...)
}
\arguments{
\item{...}{Arguments passed to \code{data.table()}}
}
\description{
\code{tidytable()} constructs a data.table, but one with nice printing features.
As such it can be used exactly like a data.table would be used.
}
\examples{
tidytable(x = 1:3, y = c("a", "a", "b"))
}
