% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytable.R
\name{tidytable}
\alias{tidytable}
\title{Build a data.table/tidytable}
\usage{
tidytable(
  ...,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)
}
\arguments{
\item{...}{Arguments passed to \code{data.table()}}

\item{.name_repair}{Treatment of duplicate names. See \code{?vctrs::vec_as_names} for options/details.}
}
\description{
\code{tidytable()} constructs a data.table, but one with nice printing features.
}
\examples{
tidytable(x = 1:3, y = c("a", "a", "b"))
}
