% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytable.R
\name{tidytable}
\alias{tidytable}
\title{Build a data.table/tidytable}
\usage{
tidytable(..., .name_repair = "unique")
}
\arguments{
\item{...}{A set of name-value pairs}

\item{.name_repair}{Treatment of duplicate names. See \code{?vctrs::vec_as_names} for options/details.}
}
\description{
Constructs a data.table, but one with nice printing features.
}
\examples{
tidytable(x = 1:3, y = c("a", "a", "b"))
}
