% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{unnest.}
\alias{unnest.}
\alias{dt_unnest_legacy}
\title{Unnest a nested data.table}
\usage{
unnest.(.data, col = NULL)

dt_unnest_legacy(.data, col = NULL)
}
\arguments{
\item{.data}{A nested data.table}

\item{col}{The column to unnest}
}
\description{
Unnest a nested data.table.
}
\examples{
nested_df <- data.table::data.table(
  a = 1:10,
  b = 11:20,
  c = c(rep("a", 6), rep("b", 4)),
  d = c(rep("a", 4), rep("b", 6))) \%>\%
  nest_by.(c, d)

nested_df \%>\%
  unnest.(data)
}
